// src/types/protoRegistry.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSON Schema file,
 * and run `pnpm run updateSchema` to regenerate this file.
 */

/**
 * Information about an author, either their name, or an object of additional fields.
 */
export type PluginAuthor = string | PluginPerson;
/**
 * Format of the plugin.
 */
export type PluginFormat = "json" | "toml" | "wasm" | "yaml";
export type Id = string;
/**
 * Strategies and protocols for locating plugins.
 */
export type PluginLocator = string;

/**
 * A collection of plugins that can be utilized by consumers.
 */
export interface PluginRegistryDocument {
	/**
	 * Path to a JSON schema.
	 */
	$schema: string;
	/**
	 * List of available plugins.
	 */
	plugins: PluginEntry[];
	/**
	 * Current version of the registry document.
	 */
	version: number;
}
/**
 * Information about a plugin.
 */
export interface PluginEntry {
	/**
	 * Information about the author.
	 */
	author: PluginAuthor;
	/**
	 * List of binary/shim names that are provided by this plugin.
	 */
	bins: string[];
	/**
	 * Description of the tool in which the plugin is providing.
	 */
	description: string;
	/**
	 * List of sources in which versions are detected from.
	 */
	detectionSources?: PluginDetectionSource[];
	/**
	 * Devicon (https://devicon.dev) for the tool.
	 */
	devicon?: string | null;
	/**
	 * Format of the plugin: WASM, or TOML
	 */
	format: PluginFormat;
	/**
	 * List of directories in which the plugin locates globally installed binaries/packages. Supports environment variables.
	 */
	globalsDirs?: string[];
	/**
	 * URL to the tool's homepage or documentation.
	 */
	homepageUrl?: string | null;
	/**
	 * Suggested identifier of the plugin. This will become the primary binary/shim name, as well as the name used on the command line, and within environment variables.
	 */
	id: Id;
	/**
	 * The location in which to acquire the plugin. More information: https://moonrepo.dev/docs/guides/wasm-plugins#configuring-plugin-locations
	 */
	locator: PluginLocator;
	/**
	 * Human readable name of the tool.
	 */
	name: string;
	/**
	 * URL to the plugin's repository.
	 */
	repositoryUrl?: string | null;
}
/**
 * Information about a person.
 */
export interface PluginPerson {
	email?: string | null;
	name: string;
	url?: string | null;
}
/**
 * A file source where the plugin attempts to detect a version from.
 */
export interface PluginDetectionSource {
	file: string;
	label?: string | null;
	url?: string | null;
}